requirements.txt
flask
openai
requests
python-dotenv
cors
flask-cors
flask-restful